+++ pll/avxLikelihood.c
@@ -42,7 +42,7 @@
 #include <stdint.h>
 
 #if defined(__ARM_NEON)
-#include "sse2neon.h"
+#include "utils/sse2neon.h"
 #else
 #include <xmmintrin.h>
 #include <pmmintrin.h>
