+++ pll/evaluatePartialGenericSpecial.c
@@ -43,7 +43,7 @@
 
 #ifdef __SSE3
 #if defined(__ARM_NEON)
-#include "sse2neon.h"
+#include "utils/sse2neon.h"
 #else
 #include <xmmintrin.h>
 #include <pmmintrin.h>
