+++ pll/makenewzGenericSpecial.c
@@ -42,7 +42,7 @@
 
 #ifdef __SSE3
 #if defined(__ARM_NEON)
-#include "sse2neon.h"
+#include "utils/sse2neon.h"
 #else
 #include <xmmintrin.h>
 #include <pmmintrin.h>
