+++ vectorclass/instrset.h
@@ -66,37 +66,37 @@
 #include <immintrin.h>                 // AVX
 #elif INSTRSET == 6
 #if defined(__ARM_NEON)
-#include "sse2neon.h"                   // SSE4.2
+#include "utils/sse2neon.h"                   // SSE4.2
 #else
 #include <nmmintrin.h>                 // SSE4.2
 #endif
 #elif INSTRSET == 5
 #if defined(__ARM_NEON)
-#include "sse2neon.h"                   // SSE4.1
+#include "utils/sse2neon.h"                   // SSE4.1
 #else
 #include <smmintrin.h>                   // SSE4.1
 #endif
 #elif INSTRSET == 4
 #if defined(__ARM_NEON)
-#include "sse2neon.h"                   // SSSE3
+#include "utils/sse2neon.h"                   // SSSE3
 #else
 #include <tmmintrin.h>                   // SSSE3
 #endif
 #elif INSTRSET == 3
 #if defined(__ARM_NEON)
-#include "sse2neon.h"                   // SSE3
+#include "utils/sse2neon.h"                   // SSE3
 #else
 #include <pmmintrin.h>                   // SSE3
 #endif
 #elif INSTRSET == 2
 #if defined(__ARM_NEON)
-#include "sse2neon.h"                   // SSE2
+#include "utils/sse2neon.h"                   // SSE2
 #else
 #include <emmintrin.h>                   // SSE2
 #endif
 #elif INSTRSET == 1
 #if defined(__ARM_NEON)
-#include "sse2neon.h"                   // SSE
+#include "utils/sse2neon.h"                   // SSE
 #else
 #include <xmmintrin.h>                   // SSE
 #endif
